\name{cutPoints}
\alias{cutPoints}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Auxiliary function for the MDLP
}
\description{
This function is required to perform the Minimum Description Length Principle.\code{mdlp}
}
\usage{
cutPoints(x, y)
}
\arguments{
  \item{x}{
a vector of numeric value
}
  \item{y}{
class variable vector
}
}
\author{
HyunJi Kim \email{polaris7867@gmail.com}
}
\seealso{
 \code{\link{cutIndex}}, 
 \code{\link{ent}}, 
 \code{\link{mergeCols}},  
 \code{\link{mdlStop}},  
 \code{\link{mylog}}, 
 \code{\link{mdlp}} .
}