\name{discretization-package}
\alias{discretization-package}
\alias{discretization}
\docType{package}
\title{
Data preprocessing, discretization for classification.
}
\description{
This package is a collection of supervised discretization algorithms. It can also be grouped in terms of top-down or bottom-up, implementing the discretization algorithms.
}
\details{
\tabular{ll}{
Package: \tab discretization\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-12-02\cr
License: GPL
LazyLoad: \tab yes\cr
}

}
\author{
Maintainer: Kim, H. J. <polaris7867@gmail.comt>
}
\references{
Chmielewski, M. R. and Grzymala-Busse, J. W. (1996). Global Discretization of Continuous Attributes as Preprocessing for Machine Learning, \emph{International journal of approximate reasoning}, \bold{Vol. 15, No. 4}, 319--331.

Fayyad, U. M. and Irani, K. B.(1993). Multi-interval discretization of continuous-valued attributes for classification learning, \emph{Artificial intelligence}, \bold{13}, 1022--1027.

Gonzalez-Abril, L., Cuberos, F. J., Velasco, F. and Ortega, J. A. (2009), Ameva: An autonomous discretization algorithm,\emph{Expert Systems with Applications}, \bold{36}, 5327--5332.

Kerber, R. (1992). ChiMerge : Discretization of numeric attributes, \emph{In Proceedings of the Tenth National Conference on Artificial Intelligence}, 123--128.

Kurgan, L. A. and Cios, K. J. (2004). CAIM Discretization Algorithm, \emph{IEEE Transactions on knowledge and data engineering}, \bold{16}, 145-153.

Liu, H. and Setiono, R. (1995). Chi2: Feature selection and discretization of numeric attributes, \emph{Tools with Artificial Intelligence}, 388--391.

Liu, H. and Setiono, R. (1997). Feature selection and discretization, \emph{IEEE  transactions on knowledge and data engineering}, \bold{9}, 642--645.

Pawlak, Z. (1982). Rough Sets, \emph{International Journal of Computer and Information Sciences}, \bold{vol.11, No.5}, 341--356.

Su, C. T. and Hsu, J. H. (2005). An Extended Chi2 Algorithm for Discretization of  Real Value Attributes, \emph{IEEE transactions on knowledge and data engineering}, \bold{17}, 437--441.

Tay, F. E. H. and Shen, L. (2002). Modified Chi2 Algorithm for Discretization, \emph{IEEE  Transactions on knowledge and data engineering}, \bold{14}, 666--670.

Tsai, C. J., Lee, C. I. and Yang, W. P. (2008). A discretization algorithm based on Class-Attribute Contingency Coefficient, \emph{Information Sciences}, \bold{178}, 714--731.

Ziarko, W. (1993). Variable Precision Rough Set Model, \emph{Journal of computer and system sciences}, \bold{Vol. 46, No. 1}, 39--59.

}
\keyword{ package }
