\name{ent}
\alias{ent}
\title{
Auxiliary function for the MDLP
}
\description{
This function is required to perform the Minimum Description Length Principle.\code{mdlp}
}
\usage{
ent(y)
}
\arguments{
  \item{y}{
class variable vector
}
}
\author{
Kim, H. J. \email{polaris7867@gmail.com}
}
\seealso{
 \code{\link{cutPoints}}, 
 \code{\link{ent}}, 
 \code{\link{mergeCols}},  
 \code{\link{mdlStop}},  
 \code{\link{mylog}}, 
 \code{\link{mdlp}} .
}