\name{insert}
\alias{insert}
\title{
Auxiliary function for Top-down discretization
}
\description{
This function is required to perform the \code{disc.Topdown()}.
}
\usage{
insert(x, a)
}
\arguments{
  \item{x}{
cut-point
}
  \item{a}{
a vector of minimum, maximum value
}
}

\author{
HyunJi Kim \email{polaris7867@gmail.com}
}
\seealso{
\code{\link{topdown}}, \code{\link{findBest}} and 
\code{\link{disc.Topdown}} . 
}